fix(protocol): serialize UpdateBlock network id with zigzag encoding
Signed block runtime ids must use zigzag on the wire; align barrel/chest traits with the corrected packet format.

+++ src/Orion/Block/Traits/BarrelTrait.cs
@@ -134,7 +134,7 @@ public class BarrelTrait : BlockTrait
             return;
         }
 
-        uint networkId = (uint)dimension.GetPermutation(x, y, z).NetworkId;
+        int networkId = dimension.GetPermutation(x, y, z).NetworkId;
 
         player.Send(
         new BlockActorDataPacket
